named tensors isolated + rationalizing imports
