+++ py4cast/datasets/poesy/__init__.py
@@ -11,7 +11,6 @@ from typing import Dict, List, Literal, Tuple, Union
 import numpy as np
 import torch
 import tqdm
-from torch.utils.data import DataLoader, Dataset
 
 from py4cast.datasets.access import (
     DataAccessor,
@@ -24,12 +23,9 @@ from py4cast.datasets.access import (
 )
 from py4cast.datasets.base import (
     DatasetABC,
-    DatasetInfo,
     Item,
     NamedTensor,
     Period,
-    TorchDataloaderSettings,
-    collate_fn,
     get_param_list,
 )
 from py4cast.datasets.poesy.settings import (
@@ -38,7 +34,6 @@ from py4cast.datasets.poesy.settings import (
     OROGRAPHY_FNAME,
     SCRATCH_PATH,
 )
-from py4cast.forcingutils import generate_toa_radiation_forcing, get_year_hour_forcing
 from py4cast.plots import DomainInfo
 from py4cast.settings import CACHE_DIR
 from py4cast.utils import merge_dicts
