+++ py4cast/datasets/titan/__init__.py
@@ -7,8 +7,6 @@ from functools import cached_property, lru_cache
 from pathlib import Path
 from typing import Callable, Dict, List, Literal, Tuple, Union
 
-import gif
-import matplotlib.pyplot as plt
 import numpy as np
 import torch
 import tqdm
@@ -32,7 +30,6 @@ from py4cast.datasets.titan.settings import (
     METADATA,
     SCRATCH_PATH,
 )
-from py4cast.forcingutils import generate_toa_radiation_forcing, get_year_hour_forcing
 from py4cast.plots import DomainInfo
 from py4cast.utils import merge_dicts
